requirements.txt
Flask>=3.0.0,<3.1.0
gunicorn
python-dotenv